Implement DataManagementSetup #450
- added UnitTests
- UnitTests :+1:
- fixes #450
