+++ Networking/DataRepository/AzureGateway/PartBindingFactory.cs
@@ -1,0 +1,31 @@
+﻿//___________________________________________________________________________________
+//
+//  Copyright (C) 2020, Mariusz Postol LODZ POLAND.
+//
+//  To be in touch join the community at GITTER: https://gitter.im/mpostol/OPC-UA-OOI
+//___________________________________________________________________________________
+
+using System;
+using UAOOI.Configuration.Networking.Serialization;
+using UAOOI.Networking.SemanticData;
+using UAOOI.Networking.SemanticData.DataRepository;
+
+namespace UAOOI.Networking.DataRepository.AzureGateway
+{
+  internal class PartBindingFactory : IBindingFactory
+  {
+    #region IBindingFactory
+
+    public IConsumerBinding GetConsumerBinding(string repositoryGroup, string processValueName, UATypeInfo fieldTypeInfo)
+    {
+      throw new NotImplementedException();
+    }
+
+    public IProducerBinding GetProducerBinding(string repositoryGroup, string processValueName, UATypeInfo fieldTypeInfo)
+    {
+      throw new NotImplementedException();
+    }
+
+    #endregion IBindingFactory
+  }
+}
