+++ Networking/DataRepository/AzureGatewayUnitTest/packages.config
@@ -1,6 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.DependencyValidation.Analyzers" version="0.11.0" targetFramework="net472" />
+  <package id="Castle.Core" version="4.4.1" targetFramework="net472" />
+  <package id="CommonServiceLocator" version="2.0.5" targetFramework="net472" />
+  <package id="Moq" version="4.14.5" targetFramework="net472" />
   <package id="MSTest.TestAdapter" version="2.1.2" targetFramework="net472" />
   <package id="MSTest.TestFramework" version="2.1.2" targetFramework="net472" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.7.1" targetFramework="net472" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net472" />
 </packages>
